% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandR.R
\name{addNewCohorts}
\alias{addNewCohorts}
\title{Add new cohorts}
\usage{
addNewCohorts(newCohortData, cohortData, pixelGroupMap, time,
  speciesEcoregion)
}
\arguments{
\item{newCohortData}{TODO: description needed}

\item{cohortData}{TODO: description needed}

\item{pixelGroupMap}{TODO: description needed}

\item{time}{TODO: description needed}

\item{speciesEcoregion}{TODO: description needed}
}
\value{
TODO: description needed
}
\description{
Does the following:
\enumerate{
  \item add new cohort data into \code{cohortdata};
  \item assign initial biomass and age for new cohort;
  \item assign the new pixelgroup to the pixels that have new cohort;
  \item update the pixelgroup map.
}
}
\details{
\code{newCohortData} must have the original \code{pixelgroup}, regenerated species
and \code{pixelindex}.
It also would be better if it has the columns of \code{cohortData} plus \code{pixelIndex}.
}
