% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandR.R
\name{addPixelGroup}
\alias{addPixelGroup}
\title{Add the correct pixelGroups to a pixel-cohortData object}
\usage{
addPixelGroup(pixelCohortData, maxPixelGroup)
}
\arguments{
\item{pixelCohortData}{# pixel groups are groups of identical pixels based
on speciesGroup x Age and ecoregionGroup}

\item{maxPixelGroup}{A length 1 numeric/integer indicating the current maximum pixelGroup value}
}
\description{
Add the correct pixelGroups to a pixel-cohortData object
}
\note{
This should not (yey) be used where age is an issue
}
