% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory.R
\name{availableMemory}
\alias{availableMemory}
\title{Report the estimated amount of available memory in the OS}
\usage{
availableMemory()
}
\value{
Numeric of class "object_size", so it can be reported in any units with format,
e.g., \code{format(availableMemory(), unit = "GB")}.
}
\description{
This reports the 'available' memory from a system \code{free} call.
If \code{free} is not installed on the system (e.g., non-Linux), returns \code{NULL}.
}
\seealso{
\code{man free} for description of available memory estimation.
}
