% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomassCalculation.R
\docType{methods}
\name{biomassCalculation}
\alias{biomassCalculation}
\alias{biomassCalculation,character,numeric,logical,numeric,character-method}
\alias{biomassCalculation,character,numeric,missing,numeric,character-method}
\alias{biomassCalculation,character,numeric,logical,numeric,missing-method}
\alias{biomassCalculation,character,numeric,missing,numeric,missing-method}
\alias{biomassCalculation,character,numeric,missing,missing,character-method}
\alias{biomassCalculation,character,numeric,missing,missing,missing-method}
\title{this function is to calculate aboveground biomass for canadian tree species based on DBH
or DBH and Height}
\usage{
biomassCalculation(species, DBH, includeHeight, height, equationSource)


  \S4method{biomassCalculation}{character,numeric,logical,numeric,character}(species,
  DBH, includeHeight, height, equationSource)


  \S4method{biomassCalculation}{character,numeric,missing,numeric,character}(species,
  DBH, height, equationSource)


  \S4method{biomassCalculation}{character,numeric,logical,numeric,missing}(species,
  DBH, includeHeight, height)


  \S4method{biomassCalculation}{character,numeric,missing,numeric,missing}(species,
  DBH, height)


  \S4method{biomassCalculation}{character,numeric,missing,missing,character}(species,
  DBH, equationSource)


  \S4method{biomassCalculation}{character,numeric,missing,missing,missing}(species,
  DBH)
}
\arguments{
\item{species}{Character string. The species name.}

\item{DBH}{Numeric. The tree's diameter at breast height (DBH, cm).}

\item{includeHeight}{Logical. Whether the biomass is calculated based on DBH and height.
If TURE, height must be provided.
Default \code{FALSE}}

\item{height}{Numeric. The tree's height (m).}

\item{equationSource}{Character. Determine the sources of equations.
Currently, this function has two
options, i.e., "Lambert2005" and "Ung2008".
Default \code{Lambert2005}}
}
\value{
Biomass (kg) and missedSpecies list that was not calculated.
}
\description{
this function is to calculate aboveground biomass for canadian tree species based on DBH
or DBH and Height
}
\note{
no note#'
}
\examples{
\dontrun{
 DBH <- seq(1, 100, 5)
 species <- c(rep("jack pine", 10), rep("black spruce", 10))
 species[1] <- "wrongSpecies"
 height <- seq(20, 40, length = 20)

 # without height information and and taking the equations from Lambert 2005
 biomass1 <- biomassCalculation(species = species, DBH = DBH)

 # with height information and and taking the equations from Lambert 2005
 biomass2 <- biomassCalculation(species = species, DBH = DBH,
                                includeHeight = TRUE, height = height)
}
}
\author{
Yong Luo
}
