% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{createPrjFile}
\alias{createPrjFile}
\title{Create a .prj file}
\usage{
createPrjFile(shpFile, prjFile)
}
\arguments{
\item{shpFile}{The filename of a shapefile to add .prj}

\item{prjFile}{The filename of a}
}
\description{
In cases where a shapefile is missing its associated .prj file.
}
