% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psp_clean.R
\name{dataPurification_ABMaturePSP}
\alias{dataPurification_ABMaturePSP}
\title{Clean Permanent Sampling Plot Data from Alberta.}
\usage{
dataPurification_ABMaturePSP(treeDataRaw, plotHeaderDataRaw)
}
\arguments{
\item{treeDataRaw}{individual tree data from Alberta PSP}

\item{plotHeaderDataRaw}{plot level data inc. location}
}
\description{
Clean Permanent Sampling Plot Data from Alberta.
}
\note{
Will not remove treeNumber = 9999
}
