% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psp_clean.R
\name{dataPurification_BCPSP}
\alias{dataPurification_BCPSP}
\title{Clean Permanent Sampling Plot Data from BC}
\usage{
dataPurification_BCPSP(treeDataRaw, plotHeaderDataRaw)
}
\arguments{
\item{treeDataRaw}{individual tree data from BC PSP}

\item{plotHeaderDataRaw}{plot level data inc. location}
}
\description{
Clean Permanent Sampling Plot Data from BC
}
