% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psp_clean.R
\name{dataPurification_NFIPSP}
\alias{dataPurification_NFIPSP}
\title{Clean Temporary Sampling Plot Data from SK Mistik forest management}
\usage{
dataPurification_NFIPSP(lgptreeRaw, lgpHeaderRaw, approxLocation)
}
\arguments{
\item{lgptreeRaw}{tree-level data from NFI PSP}

\item{lgpHeaderRaw}{plot-level data from NFI PSP}

\item{approxLocation}{locational data from NFI PSP}
}
\description{
Clean Temporary Sampling Plot Data from SK Mistik forest management
}
