% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psp_clean.R
\name{dataPurification_SKPSP}
\alias{dataPurification_SKPSP}
\title{Clean Permanent Sampling Plot Data from SK}
\usage{
dataPurification_SKPSP(SADataRaw, plotHeaderRaw, measureHeaderRaw,
  treeDataRaw)
}
\arguments{
\item{SADataRaw}{tree level data from SK PSP with age information}

\item{plotHeaderRaw}{plot-measure level data from SK PSP}

\item{measureHeaderRaw}{plot level data containing location from SK PSP}

\item{treeDataRaw}{tree level data from SK PSP}
}
\description{
Clean Permanent Sampling Plot Data from SK
}
