% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psp_clean.R
\name{dataPurification_SKTSP_Mistic}
\alias{dataPurification_SKTSP_Mistic}
\title{Clean Temporary Sampling Plot Data from SK Mistik forest management}
\usage{
dataPurification_SKTSP_Mistic(compiledPlotData, compiledTreeData)
}
\arguments{
\item{compiledPlotData}{plot-level data from SK TSP Mistic}

\item{compiledTreeData}{tree-level data from SK TSP Mistic}
}
\description{
Clean Temporary Sampling Plot Data from SK Mistik forest management
}
