% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psp_clean.R
\name{dataPurification_SKTSP_Mistik}
\alias{dataPurification_SKTSP_Mistik}
\title{Clean Temporary Sampling Plot Data from SK Mistik forest management}
\usage{
dataPurification_SKTSP_Mistik(compiledPlotData, compiledTreeData)
}
\arguments{
\item{compiledPlotData}{plot-level data from SK TSP Mistik}

\item{compiledTreeData}{tree-level data from SK TSP Mistik}
}
\description{
Clean Temporary Sampling Plot Data from SK Mistik forest management
}
