% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCC.R
\name{defineFlammable}
\alias{defineFlammable}
\title{Define flammability map}
\usage{
defineFlammable(LandCoverClassifiedMap = NULL, nonFlammClasses = NULL)
}
\arguments{
\item{LandCoverClassifiedMap}{A \code{Raster} that represents land cover
(e.g., Land Cover Classified map from 2005 or 2010 from the Canadian Forest Service).}

\item{nonFlammClasses}{numeric vector defining which classes in \code{LandCoverClassifiedMap}.}
}
\description{
Define flammability map
}
