% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeClusters.R
\docType{data}
\name{.NCONNECTIONS}
\alias{.NCONNECTIONS}
\title{Maximum number of cores that R can actually use}
\format{
An object of class \code{integer} of length 1.
}
\usage{
.NCONNECTIONS
}
\description{
Maximum number of cores that R can actually use
}
\note{
R hardcodes the maximum number of socket connections it can use (currently set to 128 in R 4.1).
Three of these are reserved for the main R process, so practically speaking, a user can create
\emph{at most} 125 connections e.g., when creating a cluster.
See \url{https://github.com/HenrikBengtsson/Wishlist-for-R/issues/28}.

We limit this a bit further here just in case the user already has open connections.
}
\keyword{internal}
