% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{factorValues2}
\alias{factorValues2}
\title{Faster version of \code{\link[raster]{factorValues}}}
\usage{
factorValues2(x, v, layer, att, append.names)
}
\arguments{
\item{x}{Raster* object}

\item{v}{integer cell values}

\item{layer}{integer > 0 indicating which layer to use (in a RasterStack or RasterBrick)}

\item{att}{numeric or character. Which variable(s) in the RAT table should be used. If \code{NULL}, all variables are extracted. If using a numeric, skip the first two default columns}

\item{append.names}{logical. Should names of data.frame returned by a combination of the name of the layer and the RAT variables? (can be useful for multilayer objects}
}
\description{
Faster version of \code{\link[raster]{factorValues}}
}
