% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOrUpdatePkg.R
\name{getOrUpdatePkg}
\alias{getOrUpdatePkg}
\title{Get or update packages}
\usage{
getOrUpdatePkg(p, minVer = "0")
}
\arguments{
\item{p}{character string denoting a package name}

\item{minVer}{character string denoting the minimum package version}
}
\value{
invoked for side effect of installing packages
}
\description{
Get or update packages
}
