% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{ghPkgName}
\alias{ghPkgName}
\title{Get the package name from a GitHub \code{repo/package@branch} string}
\usage{
ghPkgName(x)
}
\arguments{
\item{x}{character vector of package names}
}
\value{
a named character vector
}
\description{
Get the package name from a GitHub \code{repo/package@branch} string
}
\examples{
pkgs <- c("dplyr", "PredictiveEcology/pemisc", "PredictiveEcology/SpaDES.core@development")
ghPkgName(pkgs) ## "dplyr" "pemisc" "SpaDES.core"
}
