% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{isGitHubPkg}
\alias{isGitHubPkg}
\title{Check whether a package is one installed from GitHub}
\usage{
isGitHubPkg(x)
}
\arguments{
\item{x}{character vector of package names}
}
\value{
a named logical vector
}
\description{
Determines whether a string that may correspond to a package name
(e.g., \code{repo/package@branch}), could be a package installed from GitHub.
This is determined solely by the presence of a \code{/} in the string.
See example below.
}
\examples{
pkgs <- c("dplyr", "PredictiveEcology/pemisc", "PredictiveEcology/SpaDES.core@development")
isGitHubPkg(pkgs) ## FALSE TRUE TRUE
}
