% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{knnLoadFun}
\alias{knnLoadFun}
\title{Load kNN species data}
\usage{
knnLoadFun(speciesListIndex, spp, suffix, url, dPath, studyArea, rasterToMatch)
}
\arguments{
\item{speciesListIndex}{TODO: description needed}

\item{spp}{TODO: description needed}

\item{suffix}{TODO: description needed}

\item{url}{TODO: description needed}

\item{dPath}{TODO: description needed}

\item{studyArea}{TODO: description needed}

\item{rasterToMatch}{TODO: description needed}
}
\description{
Load kNN species data
}
