% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBCclusters.R
\name{makeIpsForClustersBoreaCloud}
\alias{makeIpsForClustersBoreaCloud}
\title{Create cluster on BorealCloud}
\usage{
makeIpsForClustersBoreaCloud(
  ipStart = "10.20.0.",
  ipEnd = c(68, 97, 189, 213, 220, 58, 106, 184, 217),
  availableCores = c(50, 50, 50, 50, 50, 50, 23, 23, 23),
  availableRAM = c(950, 500, 500, 500, 500, 500, 245, 245, 245),
  nProcess = 8,
  baseProcess = "cores",
  internalProcesses = 10,
  sizeGbEachProcess = 35,
  module = "fireSense",
  localHostEndIp = 68
)
}
\arguments{
\item{ipStart}{Boreal Cloud IP address that is identical across machines}

\item{ipEnd}{the part of IP adresses unique to each machine}

\item{availableCores}{the available cores on each machine}

\item{availableRAM}{the available RAM on each machine}

\item{nProcess}{the number of processes}

\item{baseProcess}{DESCRIPTION NEEDED}

\item{internalProcesses}{DESCRIPTION NEEDED}

\item{sizeGbEachProcess}{the size in Gb of each process}

\item{module}{DESCRIPTION NEEDED}

\item{localHostEndIp}{the address in ipEnd corresponding to local host}
}
\value{
DESCRIPTION NEEDED
}
\description{
Create cluster on BorealCloud
}
