% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeClusters.R
\name{makeOptimalCluster}
\alias{makeOptimalCluster}
\title{Create a parallel fork cluster}
\usage{
makeOptimalCluster(
  useParallel = getOption("pemisc.useParallel", FALSE),
  MBper = 500,
  maxNumClusters = parallelly::availableCores(constraints = "connections"),
  assumeHyperThreads = FALSE,
  ...
)
}
\arguments{
\item{useParallel}{Logical or numeric. If \code{FALSE}, returns NULL. If
\code{numeric}, then will return a cluster object with this
many cores, up to \code{maxNumClusters}}

\item{MBper}{Numeric. Passed to \code{memRequiredMB} in
\code{\link[=optimalClusterNum]{optimalClusterNum()}}}

\item{maxNumClusters}{Numeric or Integer. The theoretical upper limit
for number of nodes to use with the cluster.}

\item{assumeHyperThreads}{Logical. If \code{TRUE}, then it will more efficiently
divide the \code{maxNumClusters} by \code{useParallel}, so that there is a
lower number of cores used. This calculation may not be the ideal balance.
A message will indicate the change from \code{maxNumClusters}, if there is one.}

\item{...}{Passed to \code{makeForkClusterRandom}.
Only relevant for \code{iseed}.}
}
\description{
Given the size of a problem, it may not be useful to create a cluster.
This will make a fork cluster (so Linux only).
}
