% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{normalizeStack}
\alias{normalizeStack}
\title{Normalize each layer of a \code{RasterStack}}
\usage{
normalizeStack(x)
}
\arguments{
\item{x}{A \code{RasterStack} object.}
}
\description{
Rescales the values of of each \code{RasterLayer} between \verb{[0,1]}.
}
\author{
Tati Micheletti
}
