% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numActiveThreads.R
\name{numActiveThreads}
\alias{numActiveThreads}
\title{Count number of active threads}
\usage{
numActiveThreads(pattern = "--slave", minCPU = 50)
}
\arguments{
\item{pattern}{Character string that will be matched to the \code{ps} call}

\item{minCPU}{A numeric indicating what percent is the minimum to be considered
"active"}
}
\value{
A numeric of the number of active threads that match the pattern
}
\description{
This uses \code{ps -ef} so only works on unix-alikes. It will search
for the percent CPU use and select only those above 40
}
\examples{
\dontrun{
 ## Determine how many threads are used in each remote machine in a cluster
  cores = "localhost" # put other machine names here
  uniqueCores <- unique(cores)
  cl <- future::makeClusterPSOCK(uniqueCores, revtunnel = TRUE)
  clusterExport(cl, "numActiveThreads")
  out <- clusterEvalQ(cl, {
    numActiveThreads()
  })
  names(out) <- uniqueCores
  unlist(out)
  stopCluster(cl)

}
}
\author{
Eliot McIntire
}
