% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBCclusters.R
\name{optimalClusterNumGeneralized}
\alias{optimalClusterNumGeneralized}
\title{make IPs}
\usage{
optimalClusterNumGeneralized(
  memRequiredMB = 500,
  maxNumClusters = parallel::detectCores(),
  NumCoresAvailable = parallel::detectCores(),
  availMem = pemisc::availableMemory()/1e+06
)
}
\arguments{
\item{memRequiredMB}{NEEDED}

\item{maxNumClusters}{DESCRIPTION NEEDED}

\item{NumCoresAvailable}{DESCRIPTION NEEDED}

\item{availMem}{DESCRIPTION NEEDED}
}
\value{
DESCRIPTION NEEDED
}
\description{
make IPs
}
