% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{overlay.fun}
\alias{overlay.fun}
\title{Overlaying function}
\usage{
overlay.fun(SPP, HQ, LQ, hqLarger, lowQualityStack,
  outputFilenameSuffix = "overlay", destinationPath)
}
\arguments{
\item{SPP}{data.table column of species layer name}

\item{HQ}{data.table column of whether SPP is present in HQ layers}

\item{LQ}{data.table column of whether SPP is present in LQ layers}

\item{lowQualityStack}{low quality list/stack of rasters (will be used to fill NAs in highQualityStack)}

\item{outputFilenameSuffix}{file suffix to save raster if there was overlaying. Defautls to "overlay"}

\item{destinationPath}{directory for saved rasters}
}
\description{
Used internally in \code{overlayStacks}. Function to be applied to each row of a data.table containing information
of whether the species layer exists in the HQ and LQ data.
Only overlays if data exists in both layers, otherwise returns the layer with data
}
