% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{overlayStacks}
\alias{overlayStacks}
\title{Overlay layers within raster stacks}
\usage{
overlayStacks(
  highQualityStack,
  lowQualityStack,
  speciesList = NULL,
  outputFilenameSuffix = "overlay",
  destinationPath
)
}
\arguments{
\item{highQualityStack}{TODO: description needed}

\item{lowQualityStack}{TODO: description needed}

\item{speciesList}{TODO: description needed}

\item{outputFilenameSuffix}{TODO: description needed}

\item{destinationPath}{TODO: description needed}
}
\description{
TODO: description needed
}
