% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pemisc-package.R
\name{pemisc-package}
\alias{pemisc-package}
\alias{pemisc}
\title{\code{pemisc} package}
\description{
Miscellaneous utilities developed by the Predictive Ecology Group
(\url{https://predictiveecology.org}).
}
