% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgDepGraph.R
\name{pkgDepsGraph}
\alias{pkgDepsGraph}
\title{Build the pkg dependency graph}
\usage{
pkgDepsGraph(pkgs = c("LandR", "pemisc", "map", "SpaDES", "SpaDES.tools",
  "SpaDES.core", "SpaDES.addins", "SpaDES.shiny", "reproducible",
  "quickPlot"), plot.it = TRUE)
}
\arguments{
\item{pkgs}{A character vector of package names. Default is
c("LandR", "pemisc", "map", "SpaDES", "SpaDES.tools", "SpaDES.core",
"SpaDES.addins", "SpaDES.shiny", "reproducible", "quickPlot")}

\item{plot.it}{Logical. If \code{TRUE}, it will plot the igraph}
}
\value{
A list of 2: \code{dt} a data.table of the dependencies, and \code{dtGraph}
an igraph object that can be plotted with \code{plot()}
}
\description{
Uses \code{igraph} and \code{reproducible::pkgDep}
}
