% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepFireData.R
\name{prepFireCanada}
\alias{prepFireCanada}
\title{Download the National Burn Area Composite (Fires) in Canada}
\usage{
prepFireCanada(year, type = c("NBAC", "Polygon", "Point"),
  urlBase = "http://cwfis.cfs.nrcan.gc.ca/downloads/nbac/", ...)
}
\arguments{
\item{year}{Numeric, length 1. Which year, from 1986 to 2018 (currently)
to download}

\item{type}{Either "NBAC", "Polygon" or "Point" to get the
National Burn Area Composite, the Polygon or the Point datasets.}

\item{urlBase}{The url of the directory where the NBAC are stored. Default
is the currently known url. If this url becomes stale, please notify
the predictive ecology team.}

\item{...}{Additional arguments.}
}
\value{
A \code{SpatialPolygonsDataFrame} plus several downloaded files, including
the \file{.zip} archive and the extracted files.
Because it is running \code{prepInputs}, checksumming is occurring too.
}
\description{
Downloads data from CWFIS Datamart at \url{http://cwfis.cfs.nrcan.gc.ca/datamart}.
This runs \code{prepInputs} internally, so use can pass \code{studyArea} etc.
}
\examples{
\dontrun{
  # This will download 2 recent years
  library(sf)
  NBAC <- lapply(2016:2017, function(yr) a <- prepFireCanada(yr))
  Points <- prepFireCanada(yr, type = "Points", fun = "st_read")
  Polygons <- prepFireCanada(yr, type = "Polygons")
}
}
