% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepFireData.R
\name{prepFireNBAC}
\alias{prepFireNBAC}
\title{Download the National Burn Area Composite (Fires) in Canada}
\usage{
prepFireNBAC(year,
  urlBase = "http://cwfis.cfs.nrcan.gc.ca/downloads/nbac/", ...)
}
\arguments{
\item{year}{Numeric, length 1. Which year, from 1986 to 2018 (currently)
to download}

\item{urlBase}{The url of the directory where the NBAC are stored. Default
is the currently known url. If this url becomes stale, please notify
the predictive ecology team.}
}
\value{
A SpatialPolygonsDataFrame plus several downloaded files, including
the .zip archive and the extracted files. Because it is running
\code{prepInputs}, checksumming is occurring too.
}
\description{
This runs \code{prepInputs} internally, so use can pass
\code{studyArea} etc.
}
\examples{
\dontrun{
  # This will download 2 recent years
  NBAC <- lapply(2016:2017, function(yr) a <- prepFireNBAC(yr))
}
}
