% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCC.R
\name{prepInputsLCC}
\alias{prepInputsLCC}
\title{Simple prepInputs for LCC2005 or LCC2010}
\usage{
prepInputsLCC(year = 2005, destinationPath = ".", studyArea = NULL,
  rasterToMatch = NULL, filename2 = NULL, ...)
}
\arguments{
\item{year}{Numeric, either 2005 or 2010 (not yet implemented)}

\item{studyArea}{\code{SpatialPolygons*} object used for masking and possibly cropping
if no \code{rasterToMatch} is provided.
If not in same CRS, then it will be \code{spTransform}ed to
CRS of \code{x} before masking. Currently, this function will not reproject the
\code{x}. Optional in \code{postProcess}.}

\item{rasterToMatch}{Template \code{Raster*} object used for cropping (so extent should be
the extent of desired outcome) and reprojecting (including changing the
resolution and projection).
See details in \code{\link{postProcess}}.}

\item{...}{Passed to raster::crop}
}
\description{
A wrapper around prepInputs for the Canadian Land Cover Classification product(s)
}
