% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCC.R
\docType{methods}
\name{rasterToMatch,Raster-method}
\alias{rasterToMatch,Raster-method}
\title{Simple wrapper around \code{postProcess}}
\usage{
\S4method{rasterToMatch}{Raster}(x, studyArea, ...)
}
\arguments{
\item{x}{A Raster Layer with correct resolution and origin}

\item{studyArea}{A SpatialPolygon* object that will be sent to \code{postProcess}}

\item{rasterToMatch}{The raster to match in a \code{fasterize} call}
}
\value{
A RasterLayer object.
}
\description{
This creates a new raster layer, whose intention is to be used as
the \code{rasterToMatch} argument in further \code{prepInputs} calls.
}
