% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCC.R
\name{rasterToMatch}
\alias{rasterToMatch}
\title{Simple wrapper around \code{postProcess}}
\usage{
rasterToMatch(raster, studyArea, ...)
}
\arguments{
\item{raster}{A Raster Layer with correct resolution and origin}

\item{studyArea}{A SpatialPolygon* object that will be sent to \code{postProcess}}
}
\value{
A RasterLayer object.
}
\description{
This creates a new raster layer, whose intention is to be used as
the \code{rasterToMatch} argument in further \code{prepInputs} calls
}
