% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{toSentenceCase}
\alias{user}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{peutils}{\code{\link[peutils]{toSentenceCase}}, \code{\link[peutils]{user}}}
}}

