% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesTable.R
\name{getSpeciesTable}
\alias{getSpeciesTable}
\alias{prepSpeciesTable}
\title{Download and prepare a species traits table for use with LBMR module}
\usage{
getSpeciesTable(url = NULL, dPath = tempdir(), cacheTags = NULL)

prepSpeciesTable(speciesTable, speciesList, speciesLayers)
}
\arguments{
\item{url}{If NULL (the default), uses one from D. Cyr's LANDIS-II files:
\url{https://github.com/dcyr/LANDIS-II_IA_generalUseFiles/master/speciesTraits.csv}).}

\item{dPath}{The destination path.}

\item{cacheTags}{User tags to pass to \code{Cache}.}

\item{speciesTable}{A raw species traits table}

\item{speciesList}{TODO: description needed}

\item{speciesLayers}{TODO: description needed}
}
\value{
A \code{data.table} with columns ... TODO
}
\description{
TODO: add detailed description
}
\note{
This one is tailored to Canadian forests (?)
}
