% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messageDF.R
\name{termsInData}
\alias{termsInData}
\title{Extract terms in a quoted model statement}
\usage{
termsInData(model, data)
}
\arguments{
\item{model}{A quoted model statement}

\item{data}{A data.frame-like object with column names in which to match terms in
\code{model}}
}
\description{
Similar to \code{terms}, but this is used on a quoted model and
will only return unique matches in a \code{data}.
}
