% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{vegTypeMapGenerator}
\alias{vegTypeMapGenerator}
\title{Generate vegetation type map}
\usage{
vegTypeMapGenerator(species, cohortdata, pixelGroupMap,
  vegLeadingProportion)
}
\arguments{
\item{species}{A \code{data.table} of stuff}

\item{cohortdata}{A different \code{data.table} of stuff}

\item{pixelGroupMap}{A \code{raster}}

\item{vegLeadingProportion}{Numeric between 0-1.}
}
\description{
Generate vegetation type map
}
\author{
Eliot McIntire
}
