% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{writeSeedInfo}
\alias{writeSeedInfo}
\title{Write RNG seed info to file}
\usage{
writeSeedInfo(file = "seed.txt", append = FALSE)
}
\arguments{
\item{file}{Character specifying the filename (default \code{"seed.txt"}).}

\item{append}{Logical indicating whether to append to the file (default \code{FALSE}).}
}
\value{
Nothing returned. Invoked for its side-effect of writing to file.
}
\description{
Useful for debugging and ensuring reproducibility.
}
\author{
Alex Chubaty
}
