% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory.R
\name{availableMemory}
\alias{availableMemory}
\title{Report the estimated amount of available memory in the OS}
\usage{
availableMemory()
}
\value{
Numeric of class "object_size", so it can be reported in any units with format,
e.g., \code{format(availableMemory(), unit = "GB")}.
}
\description{
This reports the 'available' memory from a system call: \code{free} on Linux,
\code{vm_stat} on macOS, or \code{wmic} on Windows.
If neither is installed on the system, returns \code{NULL}.
}
\seealso{
\verb{man free} for description of available memory estimation.
}
