% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory.R
\name{availableMemory}
\alias{availableMemory}
\title{Report the estimated amount of available memory in the OS}
\usage{
availableMemory()
}
\value{
Returns a numeric, of class "object_size", so it can be reported
in any units with format, e.g.,
\code{format(availableMemory(), unit = "GB")}
}
\description{
This uses \code{system("free -lb", intern = TRUE))}, which will
only work on Linux-alikes. On if the program 'free' is not installed
on the system, this function returns \code{NULL}
}
