% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{loadkNNSpeciesLayers}
\alias{loadkNNSpeciesLayers}
\title{Function to load kNN species layers from online data repository}
\usage{
loadkNNSpeciesLayers(dPath, rasterToMatch, studyArea, speciesList = NULL,
  thresh = 1, url, cachePath, ...)
}
\arguments{
\item{dPath}{path to the data directory}

\item{rasterToMatch}{passed to \code{prepInputs}}

\item{studyArea}{passed to \code{prepInputs}}

\item{speciesList}{either a character vector of species names to download,
or a two-column matrix with the species names to download
and final names, with column names
\code{c("speciesNamesRaw", "speciesNamesEnd")}.
Should two raw species names share the same final name,
their biomass data will be considered as the "same species".}

\item{thresh}{the minimum number of pixels where the species must have
\code{biomass > 0} to be considered present in the study area.
Defaults to 1.}

\item{url}{the source url for the data, passed to \code{prepInputs}}

\item{cachePath}{path to the cache directory}

\item{...}{Additonal arguments}
}
\value{
a list of two elements: \code{speciesLayer}, a raster stack; and
        \code{speciesList}, a vector(?) of species names. TODO: verify this
}
\description{
TODO: description needed
}
