% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapOrDoCall.R
\name{makeClusterRandom}
\alias{makeClusterRandom}
\alias{makeForkClusterRandom}
\alias{makeSockClusterRandom}
\title{\code{makeForkCluster} with random seed set}
\usage{
makeClusterRandom(
  ...,
  type = "SOCK",
  iseed = NULL,
  libraries = NULL,
  objects = NULL,
  envir = parent.frame()
)

makeForkClusterRandom(..., iseed = NULL)

makeSockClusterRandom(..., iseed = NULL)
}
\arguments{
\item{...}{passed to \code{makeCluster}, e.g.,}

\item{type}{One of the supported types: see \sQuote{Details}.}

\item{iseed}{passed to \code{clusterSetRNGStream}}

\item{libraries}{A character vector of libraries to load in the SOCK cluster. This
is ignored if a "FORK" cluster}

\item{objects}{a character string of objects that are required inside the SOCK cluster.
Ignored if type != "SOCK"}

\item{envir}{Required if \code{objects} is passed. The environment where
\code{objects} are found.}
}
\description{
This will set different random seeds on the clusters (not the default)
with \code{makeForkCluster}.
It also defaults to creating a logfile with message of where it is.
}
