% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBCclusters.R
\name{makeIps}
\alias{makeIps}
\title{make IPs}
\usage{
makeIps(machines, ipStart, NP, proc, nProcess, sizeGbEachProcess)
}
\arguments{
\item{ipStart}{DESCRIPTION NEEDED}

\item{NP}{DESCRIPTION NEEDED}

\item{proc}{DESCRIPTION NEEDED}

\item{nProcess}{DESCRIPTION NEEDED}

\item{sizeGbEachProcess}{DESCRIPTION NEEDED}
}
\value{
DESCRIPTION NEEDED
}
\description{
make IPs
}
