% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeClusters.R
\name{makeIpsForNetworkCluster}
\alias{makeIpsForNetworkCluster}
\alias{makeIps}
\alias{optimalClusterNumGeneralized}
\title{Create IP addresses for network cluster}
\usage{
makeIpsForNetworkCluster(
  ipStart = "10.20.0.",
  ipEnd = c(68, 97, 189, 213, 220, 58, 106, 184, 217),
  availableCores = c(50, 50, 50, 50, 50, 50, 23, 23, 23),
  availableRAM = c(950, 500, 500, 500, 500, 500, 245, 245, 245),
  nProcess = 8,
  proc = "cores",
  internalProcesses = 10,
  sizeGbEachProcess = 35,
  localHostEndIp = 68
)

makeIps(machines, ipStart, proc, nProcess, sizeGbEachProcess)

optimalClusterNumGeneralized(
  memRequiredMB = 500,
  maxNumClusters = parallel::detectCores(),
  NumCoresAvailable = parallel::detectCores(),
  availMem = pemisc::availableMemory()/1e+06
)
}
\arguments{
\item{ipStart}{Boreal Cloud network address that is identical across machines
(i.e., the first, second, and third triplets of the IP address)}

\item{ipEnd}{the part of IP addresses unique to each host (i.e., the final triplet of the IP address)}

\item{availableCores}{the number of available threads on each machine.}

\item{availableRAM}{the available RAM on each machine in GB}

\item{nProcess}{the number of processes}

\item{proc}{one of \code{"cores"} or \code{"ram"}, describing the limiting factor of the
cluster computations}

\item{internalProcesses}{DESCRIPTION NEEDED}

\item{sizeGbEachProcess}{the size in GB of each process}

\item{localHostEndIp}{the address in \code{ipEnd} corresponding to local host}

\item{machines}{\code{data.frame} of compute node information containing the following columns:
\code{ipEnd}, \code{availableCores}, \code{availableRam}}

\item{memRequiredMB}{NEEDED}

\item{maxNumClusters}{DESCRIPTION NEEDED}

\item{NumCoresAvailable}{DESCRIPTION NEEDED}

\item{availMem}{DESCRIPTION NEEDED}
}
\value{
A vector of IP addresses associated with each machine in the network cluster.

DESCRIPTION NEEDED
}
\description{
Create IP addresses for network cluster
}
