% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapOrDoCall.R
\name{Map2}
\alias{Map2}
\title{\code{Map} and \code{parallel::clusterMap} together}
\usage{
Map2(..., cl = NULL)
}
\arguments{
\item{...}{passed to \code{Map} or \code{clusterMap}}

\item{cl}{A cluster object, passed to \code{clusterMap}}
}
\description{
This will send to Map or clusterMap, depending on whether cl is provided.
Also, passes the correct "function" arg... noting that \code{Map}
takes \code{f} and \code{clusterMap} takes \code{fun}, and if the
user is used to using \code{lapply}, that takes \code{FUN}. All
three of these will work.
}
\examples{

\dontrun{
a <- 1:5
Map2(a, function(x) x)

}
}
