% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{overlayStacks}
\alias{overlayStacks}
\title{Overlay layers within raster stacks}
\usage{
overlayStacks(highQualityStack, lowQualityStack,
  outputFilenameSuffix = "overlay", destinationPath)
}
\arguments{
\item{highQualityStack}{high quality list/stack of rasters (will be used preferencially)}

\item{lowQualityStack}{low quality list/stack of rasters (will be used to fill NAs in highQualityStack)}

\item{outputFilenameSuffix}{file suffix to save raster if there was overlaying. Defautls to "overlay"}

\item{destinationPath}{directory for saved rasters}
}
\description{
TODO: description needed
}
